+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [client] for Feature Flagging Service
-
 ### Changed
 
 ### Deprecated
@@ -20,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v1.34.0] - 2022-01-05
+
+### Added
+
+- [client] for Feature Flagging Service
 
 ## [v1.33.0] - 2021-12-22
 
@@ -549,7 +553,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.33.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.34.0...HEAD
+[v1.34.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.33.0...v1.34.0
 [v1.33.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.32.0...v1.33.0
 [v1.32.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.31.0...v1.32.0
 [v1.31.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.30.0...v1.31.0